begin 08-08 exercise:
I am very new to rust. This combined with my unfamiliarity with many of
the more universal concepts talked about in this project makes this last
exercise very difficult for me. I will work through it but it will
likely take a decent amount of time to figure out and, more importantly, understand thoroughly.